requirements.txt
# EV-QA-Framework dependencies
# Python 3.8+

# Testing & QA
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0

# Data processing & ML
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0
jinja2>=3.1.0

# Configuration
pyyaml>=6.0

# CAN bus protocol support
python-can==4.3.1

# Async & networking
aiohttp>=3.9.0
aiofiles==23.1.0

# DevOps & CI/CD
docker==6.1.3

# Code quality
black==23.7.0
flake8==6.0.0
isort==5.12.0

# Documentation
Sphinx==7.1.0

# Environment management
python-dotenv==1.0.0
